refactor(find-wallet): move List.tsx out of generic ProductTable dir
